+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes are documented in this file using the [Keep a CHANGELOG](htt
 
 ### Added
 
+* Added: Unimpaired statusbar toggles on: `[oe`, off: `]oe`, and toggle `coe`
 * Added: Unimpaired menu toggles on: `[oa`, off: `]oa`, and toggle `coa`
 * Added: Use new GitGutter API